+++ app/src/main/AndroidManifest.xml
@@ -16,6 +16,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".Info" />
+        <activity android:name=".Test"></activity>
     </application>
 
     <uses-permission android:name="android.permission.WAKE_LOCK" />
