--- app/src/main/res/layout/activity_value.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Value">
-
-    <TextView
-        android:id="@+id/sensorName"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textSize="20sp"
-        android:textColor="@color/black"
-        android:text="sth"
-        android:gravity="center"
-        android:layout_alignParentTop="true"/>
-
-    <TextView
-        android:id="@+id/sensorType"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textSize="20sp"
-        android:textColor="@color/black"
-        android:text="sth"
-        android:layout_below="@id/sensorName"/>
-
-    <LinearLayout
-        android:id="@+id/sensorValues"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@id/sensorType"
-        android:orientation="vertical"
-        android:layout_above="@id/send"/>
-
-    <Button
-        android:id="@+id/send"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/send"
-        android:layout_alignParentBottom="true"/>
-
-</RelativeLayout>
