--- app/src/main/res/layout/one_value.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-    <TextView
-        android:id="@+id/sensorValue"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="sth"
-        android:textSize="20sp"
-        android:textColor="@color/black"/>
-</LinearLayout>
